fix(gateway): break plugin runtime import cycle
