+++ src/gateway/node-connect-reconcile.ts
@@ -33,16 +33,6 @@ type NodeConnectPairingReconcileResult = {
   shouldClearPendingPairings?: boolean;
 };
 
-/** Publish Computer Use metadata only after the command pair is effective for this session. */
-export function resolveEffectiveComputerUseDescriptor(params: {
-  commands: readonly string[];
-  declared?: ComputerUseCapabilityDescriptor;
-}): ComputerUseCapabilityDescriptor | undefined {
-  return params.commands.includes("computer.act") && params.commands.includes("screen.snapshot")
-    ? params.declared
-    : undefined;
-}
-
 function resolveApprovedReconnectCommands(params: {
   pairedCommands: readonly string[] | undefined;
   allowlist: Set<string>;
